Create Customer Billing System Project
Objective : 
 To generate an application that can be used in any departmental store, shops, cafes, etc. for billing to the customer. This application can also be used to keep the records such as name, address, mobile number, paid amount, due amount, payment date, etc. of your regular customer. Moreover, if you have a new customer, you can add and edit the account at any time.

User-Defined Functions Used:

Although graphics has not been used in this project, the application of user-defined functions and structures have been effectively used here. The major user-defined functions used in this C project are:
void input()
void writefile()
void search()
void output()

The function void input() is used to add the new customer account i.e. with the help of these functions the parameters such as name, address, phone number, the amount paid, etc. are asked and entered. Another function void writefile() has been utilized to create a file on the hard disc of the computer for storing the information and data of a customer.
The function void search() has been used to look for previously stored accounts either by name or by the number of the customer. The fourth and the last user-defined function used in this Customer Billing System Project in C is void output() which has been defined to show the result as console output.

Features:

It can hold any number of accounts and accounts can be added to the program at any time.
The programming of simple calculations such as calculation of the due amount, balance, etc. has been embed in the code of the project.
The Customer Billing System project in C gives you the facility of searching the account in two ways, either by name of the customer or by the number of customers.
The due amount to be paid is shown as a negative balance.
If you have nothing to do with the program, you can directly exit from the main menu.